Fixed inproper resource referencing, warning instead of aborting for unordinal configs

+++ src/icmtoolchain/utils.py
@@ -83,9 +83,13 @@ def move_tree(source: str, destination: str) -> None:
 
 def copy_directory(source: str, destination: str, clear_destination: bool = False, replacement: bool = True, ignore_list: Optional[Iterable[str]] = None, relative_path: Optional[str] = None) -> None:
 	# XXX: Replace with shutil.copytree?
+	if destination.startswith(source):
+		raise ValueError(f"Destination {destination!r} cannot be copied into {source!r}!")
+
 	if clear_destination:
 		remove_tree(destination)
 	ensure_directory(destination)
+
 	for filename in os.listdir(source):
 		relative_file = join(relative_path, filename) if relative_path else filename
 		input = join(source, filename)
@@ -94,18 +98,18 @@ def copy_directory(source: str, destination: str, clear_destination: bool = Fals
 				continue
 			if input in ignore_list:
 				continue
+
 		output = join(destination, filename)
 		if isfile(input) and exists(output) and not replacement:
 			continue
+
 		if isdir(input):
 			if isfile(output):
 				os.remove(output)
 			copy_directory(input, output, clear_destination, replacement, ignore_list, relative_file)
 		elif not isdir(output):
-			try:
-				shutil.copy2(input, output)
-			except shutil.SameFileError:
-				pass
+			try: shutil.copy2(input, output)
+			except shutil.SameFileError: pass
 
 def merge_directory(source: str, destination: str, accept_squash: bool = True, ignore_list: Optional[Iterable[str]] = None, only_parent_ignore: bool = False, accept_replace: bool = True) -> None:
 	"""
